Redid the Living room walls and floor
The tutorial caused issues to the mesh that made it so loop cuts weren't lining up with both sides of the mesh. So I redid the walls making them a little thicker and doing it in a way that will make the loop cuts work. The material was added back and the old walls deleted.