Fixed up some organization and Renders
Fixed some scenes that got put in the wrong spot on my computer but now are in the right place. I also rendered some more images to show the process of texturing that will be added to the art station. I also did and updated render of the Bottled World and plan to do that same thing where I show the process of getting textures for Art Station.