Separated the Keyboard and Keys
Finished making the Keyboard and individual keys their own objects. Each key has the right name and is its own object. They are also grouped by row. Sadly the Windows/Command Key ended up on the wrong side of the space bar and it will be far to much work to fix.